Remove an outdated special case.
[Reviewed by Sung.]

Early on when we were working with sublocales, we had some confusion
regarding whether a locale with no further subdivision should be thought
of as having no sublocales, or one.  We settled on none, and put a
special case in qthreads chpl_task_getNumSublocales() to force this.
But that's actually wrong: we should return one in this case.  As of
this change, we do.


git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@22810 3a8e244f-b0f2-452b-bcba-4c88e055c3ca