Fix invalid read error in nearestMutualParentClass.chpl (--no-local).  [reviewed by bradc]
This problem surfaced after the new copyPropagation implementation was
committed (r22796).  The nature of the problem was that the program was reading
garbage in a virtual method that returns a wide string while the caller claimed
that it was returning a narrow string.

The deep reason for the failure is that in the GCC ABI, apparently space is
allocated on the stack to store the return value (if it is not of a fundamental
type).  Depending on the size of the return value, the position of other
arguments passed on the stack will vary.  Since the caller claimed the return
value was a narrow string and the callee expected it to be wide, the arguments
didn't line up and we read junk instead.

The immediate reason for the failure is that the new implementation does a
better job of propagating literals, even going so far as to replace the return
value variable with a literal in the final PRIM_RETURN primitive.  This is
advantageous in many places, so the main behavior was left in place.  Instead,
special case code was added to recognize when the symbol being replaced is the
argument of a PRIM_RETURN and its type is string.  (Copy propagation runs
before pointers are widened, so there is no distinction between wide and narrow
strings at that point.)

The fix is to prevent substitution of a literal for the return value variable in
the 'return' primitive when the return type is "string".  This forces the
return value variable to remain alive, and that allows the variable itself and
the return type of the function to be widened by insertWideReferences.

Arguably, the problem really lies in the insertWidePointers pass, since it
should recognize that a string value behaves like a pointer and therefore even
a string literal should be widened.  However, it is anticipated that the
problem will go away when the current internal string representation is
replaced by a record.  Then it will definitely be recognized as a pointer, and
no special-casing will be required.

[tested on / (--no-local)]


git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@22811 3a8e244f-b0f2-452b-bcba-4c88e055c3ca