Fix a regression in jacobi-no-inline due to r22796. [reviewed by bradc]
The fault is an assertion failure in FnSymbol::verify().
The assertion is that each method contains a declaration for its _this symbol (e.g. as an ArgSymbol).
After running the pass scalarReplace, this assertion becomes false for the method _construct__tuple (and possibly others).
The influence of the new copyPropagation implementation (r22796) is that (in contrast to the previous implementation) it allows argument symbols to substitute for other symbols in the body of a function.

In particular, in the compiler-supplied default-initializer, the "meme" argument can now substitute for uses of "this" within the code.  Thus,
    var _this;
    _this = meme;
    foo(this);
becomes
    foo(meme);
That is, after "meme" is substituted for "this", both the assignment to "this" and its declaration become dead code and are removed.

However, the way the AST is currently defined, a function is identified as a method by testing the "_this" field in its FnSymbol.  NULL means it is an unbound function, while non-NULL means it is a method.  Other uses of the AST expect the "_this" to point to a "this" symbol whose declaration lies within the scope of that function.  So even when "this" is essentially dead, we have to prevent is removal.

[tested on / (--no-local)]


git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@22812 3a8e244f-b0f2-452b-bcba-4c88e055c3ca