Change the wording for when compiler performance keys are matched
[not reviewed, trivial] 

Brad noticed that compiler performance runs had ~8k successes instead of the
usual ~4k. The reason was that compiler performance testing would print out
"success matching ..." on successful matches which is the keyword that
start_test looks for when counting the number of successes. To avoid this, this
patch just changes the wording from success matching to success finding. This
only affects compiler performance runs and not regular performance runs. 


git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@22816 3a8e244f-b0f2-452b-bcba-4c88e055c3ca