Revert the test changes in r22744. [reviewed by mferguson]
Back out the changes in tests
  remote-record-write-1
  remote-tuple-write-1
  remote-tuple-write-small-1
due to 22744, and instead update the .good files.

r22744 was intended to address performance regressions resulting from the
adoption of the new signature for the assignment operator (r22674).  Those
performance regressions were due to the fact that the "old-style" assignment
went ahead and performed a field-by-field copy (of records and (by subsumption)
tuples) by default -- as required by the specification.  The result was then
copied to the destination by the automatically-added MOVE primitive -- a
workaround used to implement assignment before the "ref" intent was available.

In the old implementation, the field-by-field assignment was performed locally
and the remote write was done in bulk, so the field-by-field copy did not
create additional network traffic.  In the new implementation, the
field-by-field update spans locales, so communication counts are higher.

The quick-fix solution was to write custom implementations of the assignment
operators for the record and tuples used in the affected tests.  But this is
just a manual implementation of an optimization that should be performed
automatically.  Accepting the new behavior should work as a goad to implement
this optimization.

For completeness, the optimization is described here:
  When copying a record, the element-by-element copy can be replaced by a bulk
  copy, provided that such a replacement does not change the semantics of the
  operation.
  The compiler knows the semantics of the assignment operators it generates
  automatically, but does not know the semantics of user-supplied assignment
  operators.  The compiler-generated assignment operators implement the
  required field-by-field copy, which when applied recursively devolves to a
  bitwise (bulk) copy.  Therefore, a compiler-generated assignment can be
  replaced by a bulk copy, but only if none of the field-wise assignments uses
  a user-defined assignment operator (recursively).

  The algorithm suggested by the above discussion is:
    For each call to an assignment operator:
      Get the function being called
      If the function is user-defined:
        continue
      If each of the field-wise assignments is compiler-defined (recursively):
        replace the assignment call with a 'move' primitive



git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@22818 3a8e244f-b0f2-452b-bcba-4c88e055c3ca