Make "is tuple type" compiler primitive [suggested by Brad, very straight-forward]
The module method isTupleType(type t) previously took the type, created a var
using it, and then checked if that var was a tuple.  This is unnecessary and
would have led to some infinite recursion in my initialization work.  To combat
this, I created a compiler check similar to the "is star tuple type" primitive
which utilized FLAG_TUPLE's existence.

Tested over std.


git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@22821 3a8e244f-b0f2-452b-bcba-4c88e055c3ca