Make the polling architecture more symmetric and simple.
[Reviewed by David.]

This rearranges polling in the gasnet comm layer so that it happens the
same way on all locales, with respect to tasking.  It used to be the
case that on locale 0 we polled in a task or thread acquired from the
tasking layer, while on other locales we polled in the main process
associated with the locale.  To take into account the load imposed by
communications polling, tasking layers had to have implicit knowledge of
this difference between locales.  Now gasnet does polling in a task or
thread acquired from the tasking layer on all locales.  The tasking
layer logic can therefore depend on explicit information, such as the
value returned by chpl_comm_numPollingTasks(), in deciding how to deal
with any polling load.

The change to call the tasking layer to start a polling task on all
locales is quite simple.  However, there is some added complication in
chpl_comm_pre_task_exit(), where we do the barrier across locales at the
end of the user code.  On the non-0 locales we cannot just call a
barrier function.  The gasnet barrier functions put callers to work
polling, and the callers of chpl_comm_pre_task_exit() on non-0 locales
are the locale processes.  They arrive in chpl_comm_pre_task_exit()
quite early, basically just as locale 0 is invoking the user code.  If
they did simple barrier calls they would thus compete for polling work
with the polling tasks or threads we just set up.  Therefore we go to a
little extra work on the non-0 locales, and set up a slow spin wait so
that we satisfy the barrier in a timely way while not interfering with
the polling.

While here, I also added an internal barrier function implemented in
terms of gasnet's barrier identifiers.  Barriers with identifiers will
allow catching certain barrier mismatches, which seems like a nice thing
to be able to do (and helped find a problem in an early version of this
mod).  I'll adapt more code to use these in the future; for now, so as
not to confuse the purpose of this commit, I'm just using this kind of
barrier in one instance.


git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@22823 3a8e244f-b0f2-452b-bcba-4c88e055c3ca