Add performance test verification
[reviewed by lydia & tom]

This patch adds the ability to verify that performance tests are working as
expected.

It adds two special 'perfkeys': verify and reject. When the regex in a verify
line matches, it is counted as being a success. A match in a reject line is
counted as a failure. If multiple verify or reject lines are in a perf keys
file the results are logically and'ed together.

verify: <regex>
verify:lineno: <regex>
verify:-lineno: <regex>

reject: <regex>
reject:lineno: <regex>
reject:-lineno: <regex>

lineno specifies what line of the output the check should match against.  If
lineno is negative the lines are counted from the end of the file to the
beginning. If no line number is given the match can be on any line.

Example line numbering of a file:
1:-3: 1 2 3 4 5
2:-2: 6 7 8 9 0
3:-1: FAILURE

When the test fails, no data is recorded and an error is emitted for the logs.


git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@22825 3a8e244f-b0f2-452b-bcba-4c88e055c3ca