Correct alias and def/use analysis for *SVEC* primitives
[reviewed by tom] 

r22606 fixed a bug where defs of aliases to records fields were not identified
as defs of the record itself. This patch is related as it fixes the same case
for homogenous tuples (which are implemented as records.) Defs to aliases of
tuple's fields were not being identified as defs of the tuple itself. The
reason for this is because of tuples being deffed/used through
PRIM_GET_SVEC_MEMBER and PRIM_SET_SVEC_MEMBER. After the pass
returnStarTuplesByRefArgs the PRIM_GET_MEMBER is turned into
PRIM_GET_SVEC_MEMBER. This turns something like PRIM_GET_MEMBER(record, x1)
where record and x1 are symExpr into PRIM_GET_SVEC_MEMBER(record, 1) where
record is still a symExpr but 1 is an integer literal representing that the
field is at an offset of one. 

This meant that for alias analysis something like 

callExpr = PRIM_GET_SVEC_MEMBER (record, 1)

would report that callExpr aliases 1 instead of the reality that callExpr
aliases the records first field. 

This patch updates alias analysis to handle this correctly as well updates the
def/use section so that PRIM_SVEC_SET_MEMBER identifies the def of the actual
field as well. 

This bug existed mostly because I had no clue that the svec primitives second
argument was an integer literal instead of the actual field. I'm kind of
surprised this wasn't an issue previously. 

A helper function was added to get the symbol of the field from a svec
primitive call. It might be useful to move that to a more central location in
the future. 



git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@22826 3a8e244f-b0f2-452b-bcba-4c88e055c3ca