Updating futures for dynamic dispatch on iterators
Updating both of these futures with a bit more information.  Dynamic
dispatch on iterators seems to work when a variable has inferred type
and the initializing expression has a clear type.  For example, in
jturner's future, removing the static type declaration of the parent
class type causes the test to work as expected.  Vass's case shows
that removing the declared type is not sufficient though (I tested
this by moving the conditional assignment into the initializer and
removing the static type -- the test still failed in that case).

Interestingly, it's not quite as naive as the variable needing to
have a single runtime type because if I add a GrandChild class to
jturner's test and re-assign the variable to a new GrandChild(),
I get the proper invocation of the grandchild's iterator.



git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@22827 3a8e244f-b0f2-452b-bcba-4c88e055c3ca