Make distribution assignment use the new signature. [reviewed by kyleb]
Upgrade assignment for distributions to use the new signature:
 - LHS has ref intent
 - return type is void.

The assignment contains two functional cases.  If the implementation (_value)
in the record-wrapped type is nil, then it is initialized by creating a new
copy of the RHS.  Otherwise, the reference count of the RHS is incremented.
In either case, the value of a is reprivatized.

In the first case, the new signature requires that a be updated directly,
rather than a new distribution object being created and returned.  I first
tried to use PRIM_ASSIGN for this purpose, but ref intent is discarded from the
arguments of RWTs when they are unwrapped.  

So I used PRIM_MOVE instead.  PRIM_MOVE handles the case of baseType -->
baseType copies in codegen, while PRIM_ASSIGN only handles the baseType -->
ref(baseType) case.  Adding other levels of indirection to the codegen for
PRIM_ASSIGN is left as pending task.

Once that is done, since RWTs behave as if they are references, there is no
harm in removing the return statement.  This brings the signature of 
  proc =(ref a: _distribution, b: _distribution) 
into line with the new standard.

[tested on / (std)]
[tested on domains/ distributions/ multilocale/ performance/ (gasnet)]


git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@22829 3a8e244f-b0f2-452b-bcba-4c88e055c3ca