Upgrade tuple assignment to the new signature. [reviewed by ]
Following r22826, it is possible to remove the return statement from 
   inline proc =(ref x: _tuple, y: _tuple)
without evoking any regressions.

Prior to that commit, alias tracking failed to recognizing that tuple elements
were being updated, because the second argument to a member access in a star
tuple is an integer (field index) rather than a field name.  The fault in alias
tracking allowed tuple assignments to be hoisted out of a loop, leading to
incorrect results.  However, this fault was only revealed after the return
statement was removed from tuple assignment, because the compiler-inserted move
(for assignments returning a value) formerly prevented that hoisting.

The removal of the return statement results in one fewer get and put per cycle
in the performance tests remote-tuple-write-1 and remote-tuple-write-small-1,
so the .good files were updated to reflect this improvement.

[tested on / (std)]
[tested on domains distributions multilocale performance (gasnet-everything)]


git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@22832 3a8e244f-b0f2-452b-bcba-4c88e055c3ca