Create "is sync type" compiler primitive [reviewed by Brad]
To aid in my initialization work, I created a compiler primitive which takes a
type and returns a boolean indicating if the type provided is a sync.  I
additionally added small helper methods that check if a provided variable is a
sync, and created two tests to check both these methods.

tested over std


git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@22840 3a8e244f-b0f2-452b-bcba-4c88e055c3ca