Create "is single type" compiler primitive [trivially similar to r22840]
To aid in my initialization work, I created a compiler primitive which takes a
type and returns a boolean indicating if the type provided is a single.  I
additionally added small helper methods that check if a provided variable is a
single, and created one test to check both these methods.

tested over examples


git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@22841 3a8e244f-b0f2-452b-bcba-4c88e055c3ca