Add a heuristic optimization to replace trivial record assignments with verbatim copies. [reviewed by vass]
We define a trivial assignment as one which can be replaced by a verbatim
(bulk) copy.  A verbatim (bitwise) copy is performed by the PRIM_ASSIGN and
PRIM_MOVE primitives.

Basically, given a record like
  record R { var x, y, z: int; }
if assignment for it looks like
  proc =(ref lhs:R, rhs:R) {
    lhs.x = rhs.x;
    lhs.y = rhs.y;
    lhs.z = rhs.z;
  }
then the whole function can be potentially be replaced by
  proc =(ref lhs:R, rhs:R) { __primitive("=", lhs, rhs); }
.  The tricky part is that in order for the assignment to be trivial, the
assignments implementing the field-by-field copies must also be trivial -- all
the way down to fundamental types.  

For an assignment function to be trivial, it must be defined as a
field-by-field assignment covering all the fields defined in the type being
copied, and this property must hold recursively on each of the field-wise
assignments.  Since the field-wise assignment involves field setters and
getters, and it was deemed difficult to determine whether a function is one
such (under the desire to keep contained the AST rework required to support
this optimization), a heuristic was employed instead: Using the assumption that
when the "compiler generated" flag is applied to an assignment, the assignment
is either trivial already or contains only field-wise assignments (covering all
fields), we only need to check that triviality holds recursively.

This simplification relies on the "compiler generated" flag being applied only
to assignment implementations that meet the above criteria.  It does not check
explicitly to ensure that only getter, setter and assignment functions are
called, nor that all the fields in a given record type are covered.  Should the
AST be reworked such that setter and getter functions were labelled
appropriately, then adding rigor to this optimization would be straightforward.

We run the recursive test as we examine each function at the top level, but
keep track of functions as they are visited to ensure that we visit each
function just once.  We use the FLAG_TRIVIAL_ASSIGNMENT to store the result of
a previous visit.  Some trivial assignment functions in the modules are marked
explicitly using FLAG_TRIVIAL_ASSIGNMENT in its pragma form.  This is to
indicate to this optimization pass that their bodies contain only a PRIM_MOVE
or PRIM_ASSIGN.  We at present do not recognize such functions for simplicity;
instead we expect that they are marked with the corresponding pragma in the
source code.

The actual test that we use to determine if an assignment function is trivial
is:
  - It is marked as "compiler generated".
  - Its return type is "void".
  - The base types of its left and right operands match.
  - Each of the primitives appearing in the body of the function is of an
    expected kind.
  - Each function called is either not an assignment function or is trivial.

Note that even if an assignment function is already defined as a verbatim copy,
the body is replaced with a new verbatim copy.  This is an expediency.  It
avoids having to test if the body is a verbatim copy (which would render the
"trivial assignment" pragma unnecessary).

This optimization is expected to improve scalar performance and reduce code
size.  In addition, should the lhs and rhs variables straddle a node boundary,
the replacement of a field-wise copy with a verbatim copy will result in
reduced comm counts and is expected to result in faster communication overall
(due to reduced communication overhead).

This optimization works for tuples and unions as well.  Assignment beween
variables of class type is predefined to be a pointer copy, so this
optimization will not affect class assignment.


DETAILS

Index: test/performance/ferguson/remote-record-write-1.good
Due to the optimization, the put count dropped from 300000 to 1000000, so the
.good file was updated accordingly.

Index: modules/standard/SysBasic.chpl
The assignment of syserr objects is explicitly marked as trivial.

Index: modules/internal/ChapelBase.chpl
Trivial assignments in ChapelBase are marked accordingly.  This will allow
compiler generated assignments involving records of these types to be marked as
trivial.

Index: compiler/include/flags_list.h
Added the new "trivial assignment" flag.

Index: compiler/include/checks.h
Index: compiler/include/passes.h
Index: compiler/include/log.h
Index: compiler/main/checks.cpp
Index: compiler/main/passlist.h
Index: compiler/optimizations/bulkCopyRecords.cpp
Index: compiler/optimizations/Makefile.share
Added the new bulkCopyRecords pass.

Index: compiler/passes/buildDefaultFunctions.cpp
Added the "assignop" flag to compiler-generated assignment functions.  This
flag is used as a relatively cheap filter, to focus attention on functions
likely to be named "=" before performing the more-expensive strcmp test.

Index: compiler/resolution/wrappers.cpp
The "assignop" flag is also propagated to wrapper functions, so the test for
whether an assignment is trivial descends through wrapper functions.  Coercion
wrappers are expected to conform to the pattern of a trivial assignment, and if
so may be replaced by a verbatim copy.


[tested on / (std)]
[tested on performance/ferguson/ (gasnet-everything)]


git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@22845 3a8e244f-b0f2-452b-bcba-4c88e055c3ca