Modify threads-pthreads to possibly use __thread for TLS
[reviewed by Greg]

In an approach taken from qthreads, we modify the threads-pthreads
implementation to use __thread when it is available by defining
macros for declaring, initializing, deleting, reading, and writing
thread local storage. When __thread is not available, we use
the usual pthread_key_create/delete pthread_getspecific/setspecific calls.

Lastly, I noticed that in pthread_func, the threadEnd function is
not called if the thread is not canceled (but the thread function
completes normally). I believe that it should be called in either
case -- it might free memory for example (and does for FIFO), so this
patch calls pthread_cleanup_pop(1) -- the 1 meaning to invoke the
cleanup handler. I don't think that this comes up with FIFO's current
use of threads-pthreads because it seems that thread_begin never exits;
nonetheless I think this is a worthwhile bug fix in case the future
usage pattern changes.



git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@22847 3a8e244f-b0f2-452b-bcba-4c88e055c3ca