Rename ClassType as AggregateType. [reviewed by Lydia]
In the compiler, the class formerly known as ClassType actually represents
classes, unions and records.  So renaming it as AggregateType (following the
C/C++ naming convention) more accurately reflects its use.

All instances of "ClassType", "isClassType" and "toClassType" were renamed to
their "AggregateType" counterparts, but the signatures of
getDataClassType() and setDataClassType() were left unchanged.

I also renamed the classTag field as "aggregateTag", renamed the ClassTag enum
as AggregateTag and renamed its members using the prefix "AGGREGATE_" in place
of "CLASS_".

[tested on examples/ (std)]


git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@22858 3a8e244f-b0f2-452b-bcba-4c88e055c3ca