Add compiler performance testing to --no-local and --fast
[reviewed by thomas]

Add compiler performance testing to --no-local(chap08) and --fast(chap12).
These are probably the only other configurations that compiler timings are
likely to differ (beyond different backend compilers, but that's all in make
binary.) comm=gasnet testing was not added, because it is almost identical to
--no-local with slight variation in make binary. Baseline timing might be
interesting if in the future if the chapel compiler optimizations are not on by
default but for now it would just be extra noise. 


git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@22859 3a8e244f-b0f2-452b-bcba-4c88e055c3ca