Use assign-to-ref for string assignment. [reviewed by bradc]
Convert string assignment to use the new signature: LHS is passed by ref and
return type is void.

__primitive("=", a, b); does not work in the body of this assignment, because
there is special handling for the assignment of wide strings.  (Not only the
descriptor, but the bytes contained are copied.)  

Note that this implementation of assignment continues to assume that local
strings are shared (so they cannot be deleted).  This version of assignment
should be replaced by a more sophisticated version when the strings-as-records
work is complete.

[tested on / (std)]
[tested on / (no-local)]


git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@22862 3a8e244f-b0f2-452b-bcba-4c88e055c3ca