Minor improvements in the compiler and modules.
[Reviewed by Elliot]

* In checkFormalActualBaseTypesMatch() and checkFormalActualTypesMatch(),
ensure with INT_FATAL that whenever the actual is 'nil', the formal
is of a (Chapel) class type.

* In createTaskFunctions(), strengthen an assertion.

* In CyclicDist.chpl, do not set the pid upon 'new CyclicArr(...)'.
Rationale: 'pid' should be set when the object is privatized using
_newPrivatizedClass. Setting 'pid' to that of another already-
privatized object breaks some unspoken invariants/assumptions.

* In ChapelArray.chpl, replace _isPrivatized(a._value) with
_isPrivatized(a._valueType), where 'a' is a _distribution.
This is for consistency with the other uses of _isPrivatized.
Also, for privatized distributions a._value-the-accessor is an int,
so I am unclear how it worked at all before this change.


git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@22865 3a8e244f-b0f2-452b-bcba-4c88e055c3ca