Fix regressions due to r22858, plus minor tweaks. [not reviewed]
Index: compiler/optimizations/loopInvariantCodeMotion.cpp
The regressions were due to a missed logical negation on line 507 of
loopInvariantCodeMotion().
[tested with test_nested_var_iterator3 (--no-scalar-replacement)]

Index: compiler/parser/chapel.ypp
I noticed that the extraction of the comment on a union declaration
(for chpldoc) was missing, and added it.
[tested with classes/jplevyak/union_type_hierarchy*.chpl]
[tested with classes/bradc/unions/]

Index: compiler/AST/type.cpp
Shortened the name of the constructor argument from initAggregateTag
to initTag.

Index: compiler/include/type.h
Changed the name of the argument to the AggregateType constructor to
"initTag", to match its definition in type.cpp.

Index: compiler/passes/parallel.cpp
Removed an extra space.

[test-built and run with hello.chpl]


git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@22868 3a8e244f-b0f2-452b-bcba-4c88e055c3ca