Distinguished in the compiler the task functions created due to 'cobegin' or 'coforall' from those due to 'begin'; reduced in the generated code the number of reference count increments/decrements for record-wrapped types and of copy constructors and destructors for records.
[reviewed by Tom]

Motivation: cobegin/coforall task functions do not outlive, at runtime,
their enclosing lexical scope, whereas begin ones can. The compiler
can benefit from this information. In particular, this commit eliminates
reference count increments and decrements upon entering and exiting
the task functions for cobegins and coforalls.


COMPILER CHANGES

* Replaced
 PRIM_BLOCK_ON_NB
with
 PRIM_BLOCK_BEGIN_ON
 PRIM_BLOCK_COBEGIN_ON
 PRIM_BLOCK_COFORALL_ON

This was one way to let the compiler distiuguish between the task functions
for 'begin' vs. for 'cobegin'/'coforall'.

I chose the new primitives to correspond to the language constructs.
Instead we could create primitives to reflect particular properties
of those language constructs, which is what PRIM_BLOCK_ON_NB did.
However, the former was conceptually simpler to me, and I am not sure
the latter would have any significant benefits. Especially given that
the primitives are around only between the parser and createTaskFunctions(),
i.e. they are gone before resolve().

The new primitives correspond to STMT in the following language constructs,
respectively:

 begin on EXPR STMT
 cobegin { ...; on EXPR STMT; ... }
 coforall ... on EXPR STMT

They are introduced by the parser as a result of the "begin+on" optimization
- where PRIM_BLOCK_ON_NB used to be introduced.

They are converted to task functions in createTaskFunctions(), just like
PRIM_BLOCK_ON_NB used to be. Now createTaskFunctions() can set different
flags on the task functions for each syntactic construct.

* So, the 'on'-related task functions now have the following flags:

 begin+on       FLAG_ON  FLAG_NON_BLOCKING  FLAG_BEGIN
 cobegin+on     FLAG_ON  FLAG_NON_BLOCKING  FLAG_COBEGIN_OR_COFORALL
 coforall+on    FLAG_ON  FLAG_NON_BLOCKING  FLAG_COBEGIN_OR_COFORALL
 on (blocking)  FLAG_ON

* FLAG_COBEGIN_OR_COFORALL and FLAG_COBEGIN_OR_COFORALL_BLOCK
mark task functions and their wrappers, respectively.

New in this commit: they mark BOTH local and 'on'-containing
task functions/wrappers.

Before they marked only the local ones.

There are a couple of cases where the compiler needs to distinguish
local vs. 'on'-related task functions. I added an explicit check for
!hasFlag(FLAG_ON_BLOCK) in those cases:
  - expr.cpp
  - insertLineNumbers.cpp

* Replaced checks/asserts/... on PRIM_BLOCK_ON_NB with the same
checks/asserts/... on the three new prims - straightforward.


* parallel.cpp - insertAutoCopyDestroyForTaskArg()

This function now inserts autoCopy/autoDestroy only for 'begin'
task functions, not for 'cobegin' or 'coforall'.

For record-wrapped types, this means no reference counter increments/
decrements upon cobegin/coforall. Those are indeed necessary only for 'begin'.

For records, this means no autoCopy (aka copy constructor) at the time
the corresponding task is created, upon cobegin/coforall. (And,
correspondingly, no matching destructor at the end of the task
function.) This is OK because the constructor and destructor are
inserted by addLocalCopiesAndWritebacks() into the task function itself.
(Meaning that the argument bundle for the task function will contain
a copy of the original record, not the autoCopy'ed one.)

For a 'begin' task function, two autoCopys and autoDestroys are
going to be inserted, just like it was the case before this commit.
One set is added by addLocalCopiesAndWritebacks() and the other set
by insertAutoCopyDestroyForTaskArg().

We could try to remove the set added by addLocalCopiesAndWritebacks(),
although it did not work for me upon a first try.

* Selected reviewer's comment: In general, I approve of the direction
of this change -- making the AST more similar to the input in the
early passes and normalizing these in later passes.  You achieve this
by adding block identifier primitives that are closely related to the
source code and then using flags to summarize the important properties
of these different block types.

* Selected reviewer's comment: It would be nice to move the code to
add constructor and destructor calls to a pass that precedes
resolution.  Also, reworking record-wrapped types so they use
constructors and destructors for reference counting "like everyone
else" would be a nice cleanup.  But of course, that is beyond the
scope of this patch.


TEST CHANGES

* I adjusted their .good for these tests, which improve fork counts,
  reverting the effect of r22438:

    modules/sungeun/init/printInitCommCounts
    studies/hpcc/CommDiags/stream-ep_commDiags

They still fail due to get count increase, as noted in REGRESSIONS
as due to r22674 - I did not change the get count in .good.

* This test also improves fork counts, reverting the effect of r22438.
It also improves get counts vs. pre-r22438 state.
I adjusted the .good with the new counts, so the test should pass now.

    studies/ssca2/performance/SSCA2_commDiags


git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@22875 3a8e244f-b0f2-452b-bcba-4c88e055c3ca