Do not add autoCopy/autoDestroy inside 'begin' task functions during resolution.
[reviewed by Tom]

autoCopy/autoDestroy for 'begin' task function arguments are
[as of my vass.commcounts.patch] added in insertAutoCopyDestroyForTaskArg().
Prior to this commit this was also done in addLocalCopiesAndWritebacks()
during resolution, which was adding an extra autoCopy/autoDestroy pair.
This change disables the latter specifically for 'begin' task functions.

So which is the right place to add autoCopy/autoDestroy -
insertAutoCopyDestroyForTaskArg or addLocalCopiesAndWritebacks ?

addLocalCopiesAndWritebacks adds autoCopy/autoDestroy inside
the (task) function. For 'begin' this is too late - by the time
autoCopy executes inside the task function, the enclosing scope may
have already exited and (e.g. for arrays) deallocated the reference-
counted object.

insertAutoCopyDestroyForTaskArg adds autoCopy so that it executes
right before the task is created. This is the desired effect.
BTW the consequence is that the value that's stored in the argument
bundle is the the result of an autoCopy, vs. for other task
and non-task functions it is the source of an autoCopy.

Both insertAutoCopyDestroyForTaskArg and addLocalCopiesAndWritebacks
put autoDestroy at the end of the task function, so there is
no difference in this regard.

We do compromise uniformity: autoCopy/autoDestroy
are created differently for begin task functions vs. for
cobegin/coforall task functions and non-tasking functions.
The reason is to make this a small change.


git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@22876 3a8e244f-b0f2-452b-bcba-4c88e055c3ca