Add test description to compiler performance graphs
[reviewed by thomas]

Previously the compiler performance title was something like: 

Chapel Compiler Performance Graphs for chap08. 

All this patch does is change that to: 

Chapel Compiler Performance Graphs for chap08 (--no-local) 

This is so that internally we can keep track of what's being tested easier
because it's easy to forget what is running on each machine and it probably
even more so important for external collaborators. This wasn't really a problem
before because we only tested compiler performance on chap01 and with r22859
added --no-local and --fast testing as well. 


git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@22884 3a8e244f-b0f2-452b-bcba-4c88e055c3ca