Improving and retiring a future
[reviewed by hilde]

In reviewing futures related to OOP/dispatch the other night, I came
across this one which was attributed to a mail I sent out in August
2012, but not captured properly.  The test now works as intended (and
probably has for some time now), but as captured, had inconsistencies
and issues that would prevent it from matching its .good.  Here, I'm
correcting those, making improvements to the test, and retiring the
future.



git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@22889 3a8e244f-b0f2-452b-bcba-4c88e055c3ca