Make BlockDom iterator ignore running tasks on the initiating task
[reviewed by Brad and David]

This patch addresses a performance issue that arose when the default
for dataParIgnoreRunningTasks was changed to false (r22641).  On the
initiating locale, the local task created by the coforall statement
can start before the initiating task is finished creating all the
tasks.  When this local task computes the number of tasks to use
locally using ignoreRunning=false, it may count the initiating task.

This commit gets around this issue in cases where the initiating task
is thought to be the only task running on the locale, by using
ignoreRunning=true in these cases.

This is a some what dissatisfactory solution, but it's the best we can
do with the current tasking interface.  Here are a couple of thoughts
on other options.

- We have talked about moving making the task counting a Chapel-level
  concept, and in that world, we could have a task request that it
  *not* be counted or provide a count that excludes such tasks.

- We have also talked about providing a vector fork operation for
  coforalls.  In that situation, the initiating task could pull off
  one local task for itself.  This has the downside of delaying
  starting of the local task, but it has the advantage of not having
  to go into the tasking layer.



git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@22892 3a8e244f-b0f2-452b-bcba-4c88e055c3ca