Optimize candidate pair removal in copyPropagation. [reviewed by eronagha]
Added back the reverse map of available pairs to optimize the invalidation of
candidate pairs.  This has the desired optimization effect (making the
algorithm faster without changing its behavior).

On the studies/lulesh/sungeun/lulesh test case where copy propagation was
taking a minute or more, the time dropped to 3-4s.  The number of emitted lines
of code remained the same.

The reverse map uses the RHS symbols as keys.  Since several LHS symbols may
have the same RHS, the mapped value is a list of LHSs associated with that
RHS.  When a new candidate pair is created, the LHS is pushed onto the list
associated with that RHS.  

When a symbol is invalidated (because its associated value is update), there
will be at most one candidate pair that contains it on the LHS.  This pair is
removed from the forward map directly.  Using the reverse map, all candidate
pairs having that symbol on the RHS are also removed.  This leaves that entry
in the reverse map empty, so it can be removed.  

Note that when the pair (lhs1, rhs1) is removed because lhs1 == sym, the entry
(rhs1, [..., lhs1, ...]) will remain in the reverse map.  That is OK, because
if rhs1 is invalidated at a later time, the attempt to remove lhs1 from the
forward map will fail silently.  This is probably cheaper than trying to keep
the reverse map precisely up-to-date.


[tested on / (std)]


git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@22893 3a8e244f-b0f2-452b-bcba-4c88e055c3ca