Better const args for denested functions and remote value forwarding
[reviewed by hilde, vass, with one slight post-review modification
that should be reviewed post-commit]

This commit improves the insertion of formals to represent non-local
variables a bit better when de-nesting functions by strengthening the
check for the const-ness of the original variable and not making the
argument type have 'ref' intent if it is constant.  This reduces the
number of unnecessary references we use for nested functions (user or
compiler-inserted).

The one bit of this commit that was not reviewed is that in the 11th
hour, I switched from using the isConstant() method to adding a new,
stricter, method called isConstValWillNotChange().  The former says
that something is constant in the sense that you cannot assign to it
(e.g., it includes 'const ref' arguments) while the latter says that
something is constant in the sense that you can rely on it not
changing out from under you (so 'const ref' is excluded).  As a
cleanup step, I think we should change isConstant() to
isConstCannotAssign() and then switch existing calls to isConstant()
to one or the other version.

The other major part of this patch is an improvement to remote value
forwarding for arguments with 'const in' intent in the presence of
sync variable accesses.  Previously some of these cases were not
remote value forwarded, but they can/should be.  This change will have
more benefit once more formal_temps are removed.  For the 'reduce*'
tests in the distribution robustness performance suite, this resulted
in one fewer get per locale.

I added an editorial note in the comments of flattenFunctions.cpp
pointing out that making a formal have a 'ref' type but 'blank'
intent, as we currently do, is arguably inconsistent, but left that as
future work (perhaps near-future work, as we've talked about pushing
the insertion of 'ref' types much closer to codegen in recent
discussions).  See the comment itself for more detail.

I also added a test that presents a simplified version of the case
that caused me to note the lack of remote value forwarding in the
instance of 'const in' arguments (co-written by Vass).  This test
works at present with or without r.v.f. but once fewer formal temps
are inserted (coming soon to a trunk near you), lack of r.v.f.  breaks
this test (and SSCA2, from which it was derived).



git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@22894 3a8e244f-b0f2-452b-bcba-4c88e055c3ca